Extract the bulk of map_test.go into new map.go
This allows the map integration tests to be run in
other environments.
